+++ libs/ui/src/lib/accounts-table/accounts-table.component.ts
@@ -66,6 +66,7 @@ export class GfAccountsTableComponent {
   public readonly showAllocationInPercentage = input<boolean>();
   public readonly showBalance = input(true);
   public readonly showFooter = input(true);
+  public readonly showQuantity = input<boolean>();
   public readonly showValue = input(true);
   public readonly showValueInBaseCurrency = input(true);
   public readonly totalBalanceInBaseCurrency = input<number>();
@@ -101,6 +102,10 @@ export class GfAccountsTableComponent {
 
     if (this.showActivitiesCount()) {
       columns.push('activitiesCount');
+    }
+
+    if (this.showQuantity()) {
+      columns.push('quantity');
     }
 
     if (this.showBalance()) {